move/create files `traverse.hpp` and `layout.hpp`
- move files to reflect namespace strcture
- add description
