+++ include/alpaka/onAcc/layout.hpp
@@ -8,21 +8,25 @@ namespace alpaka::onAcc
 {
     namespace layout
     {
+        /** Generates indices scattered based on the number of worker threads for each dimension.*/
         struct Strided
         {
         };
 
+        constexpr auto strided = Strided{};
+
+        /** Indices will be contiguous within each dimension for each worker thread. */
         struct Contigious
         {
         };
 
+        constexpr auto contigious = Contigious{};
+
+        /** The index layout will automatically selected based on the executor. */
         struct Optimized
         {
         };
 
-        constexpr auto strided = Strided{};
-        constexpr auto contigious = Contigious{};
         constexpr auto optimized = Optimized{};
-
     } // namespace layout
 } // namespace alpaka::onAcc
